mono/evals/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aipr-evals"
version = "0.1.0"
description = "Evaluation framework for AI-PR-Reviewer"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"

dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
]

[project.scripts]
aipr-eval = "scripts.run_eval:main"
aipr-report = "scripts.report:main"
